update: 시간 파싱 및 변환 로직 유틸리티 클래스로 분리 및 리팩토링
