+++ src/main/java/team/washer/server/v2/domain/reservation/service/impl/CancelOverdueReservationServiceImpl.java
@@ -16,6 +16,7 @@ import team.washer.server.v2.domain.reservation.util.PenaltyRedisUtil;
 import team.washer.server.v2.domain.smartthings.service.DetectMachineRunningService;
 import team.washer.server.v2.domain.smartthings.service.QueryDeviceStatusService;
 import team.washer.server.v2.domain.user.entity.User;
+import team.washer.server.v2.global.util.DateTimeUtil;
 
 @Slf4j
 @Service
@@ -48,7 +49,8 @@ public class CancelOverdueReservationServiceImpl implements CancelOverdueReserva
                 var isRunning = detectMachineRunningService.execute(machine.getDeviceId());
 
                 if (isRunning) {
-                    var expectedCompletionTime = getExpectedCompletionTime(machine.getDeviceId());
+                    var expectedCompletionTime = DateTimeUtil.getExpectedCompletionTime(queryDeviceStatusService,
+                            machine.getDeviceId());
                     reservation.start(expectedCompletionTime);
                     reservationRepository.save(reservation);
 
@@ -68,21 +70,5 @@ public class CancelOverdueReservationServiceImpl implements CancelOverdueReserva
                 log.error("Error processing timeout for reservation {}", reservation.getId(), e);
             }
         }
-    }
-
-    private LocalDateTime getExpectedCompletionTime(String deviceId) {
-        try {
-            var status = queryDeviceStatusService.execute(deviceId);
-            var completionTimeStr = status.getCompletionTime();
-
-            if (completionTimeStr != null && !completionTimeStr.isBlank()) {
-                var utcTime = java.time.ZonedDateTime.parse(completionTimeStr);
-                return utcTime.withZoneSameInstant(java.time.ZoneId.of("Asia/Seoul")).toLocalDateTime();
-            }
-        } catch (Exception e) {
-            log.warn("Failed to get expected completion time for device: {}", deviceId, e);
-        }
-
-        return LocalDateTime.now().plusHours(1);
     }
 }
