+++ src/main/java/team/washer/server/v2/domain/smartthings/service/impl/DetectMachineCompletionServiceImpl.java
@@ -1,8 +1,6 @@
 package team.washer.server.v2.domain.smartthings.service.impl;
 
 import java.time.LocalDateTime;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
 import java.util.Optional;
 
 import org.springframework.stereotype.Service;
@@ -12,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import team.washer.server.v2.domain.smartthings.service.DetectMachineCompletionService;
 import team.washer.server.v2.domain.smartthings.service.QueryDeviceStatusService;
+import team.washer.server.v2.global.util.DateTimeUtil;
 
 @Service
 @RequiredArgsConstructor
@@ -33,23 +32,13 @@ public class DetectMachineCompletionServiceImpl implements DetectMachineCompleti
 
             if (isFinished) {
                 log.debug("Device {} job is finished", deviceId);
-                return Optional.of(LocalDateTime.now());
+                return Optional.ofNullable(DateTimeUtil.parseAndConvertToKoreaTime(status.getCompletionTime()));
             }
 
             return Optional.empty();
         } catch (Exception e) {
             log.warn("Failed to detect completion state for device: {}", deviceId, e);
             return Optional.empty();
-        }
-    }
-
-    private LocalDateTime parseAndConvertToKoreaTime(String timeStr) {
-        try {
-            var utcTime = ZonedDateTime.parse(timeStr);
-            return utcTime.withZoneSameInstant(ZoneId.of("Asia/Seoul")).toLocalDateTime();
-        } catch (Exception e) {
-            log.warn("Failed to parse time: {}", timeStr, e);
-            return LocalDateTime.now();
         }
     }
 }
